move graph gen to submodule
